Refactored user restore bloc as more lightweight cubit
