--- lib/bloc/user_restore/user_restore_event.dart
@@ -1,10 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class UserRestoreEvent extends Equatable {
-  const UserRestoreEvent();
-}
-
-class UserRestoreStarted extends UserRestoreEvent {
-	@override
-	List<Object> get props => [];
-}
